+++ src/utils/images/imageCrop/cropSvgImage.ts
@@ -1,52 +1,4 @@
 /**
- * Utility functions for handling image cropping in different formats
- */
-
-/**
- * Crop a JPEG image
- * @param image The source image element
- * @param crop The crop dimensions {x, y, width, height}
- * @returns The cropped image data URL
- */
-export const cropJpegImage = (
-    image: HTMLImageElement,
-    crop: { x: number; y: number; width: number; height: number }
-): string => {
-    const canvas = document.createElement('canvas')
-    const ctx = canvas.getContext('2d')
-    if (!ctx) return ''
-
-    canvas.width = crop.width
-    canvas.height = crop.height
-
-    ctx.drawImage(image, crop.x, crop.y, crop.width, crop.height, 0, 0, crop.width, crop.height)
-
-    return canvas.toDataURL('image/jpeg')
-}
-
-/**
- * Crop a PNG image (preserves transparency)
- * @param image The source image element
- * @param crop The crop dimensions {x, y, width, height}
- * @returns The cropped image data URL
- */
-export const cropPngImage = (
-    image: HTMLImageElement,
-    crop: { x: number; y: number; width: number; height: number }
-): string => {
-    const canvas = document.createElement('canvas')
-    const ctx = canvas.getContext('2d')
-    if (!ctx) return ''
-
-    canvas.width = crop.width
-    canvas.height = crop.height
-
-    ctx.drawImage(image, crop.x, crop.y, crop.width, crop.height, 0, 0, crop.width, crop.height)
-
-    return canvas.toDataURL('image/png')
-}
-
-/**
  * Crop an SVG image preserving vector data
  * @param svgUrl The source SVG URL
  * @param crop The crop dimensions {x, y, width, height} - relative to the displayed image
@@ -121,42 +73,4 @@ export const cropSvgImage = async (
         console.error('Error cropping SVG:', error)
         return ''
     }
-}
-
-/**
- * Detect image type from URL or data URL
- * @param imageUrl The image URL to check
- * @returns The detected image type ('jpeg', 'png', 'svg', or 'unknown')
- */
-export const detectImageType = async (imageUrl: string): Promise<'jpeg' | 'png' | 'svg' | 'unknown'> => {
-    // Check for obvious patterns in the URL or data URL
-    if (imageUrl.startsWith('data:image/jpeg') || imageUrl.endsWith('.jpg') || imageUrl.endsWith('.jpeg')) {
-        return 'jpeg'
-    } else if (imageUrl.startsWith('data:image/png') || imageUrl.endsWith('.png')) {
-        return 'png'
-    } else if (imageUrl.startsWith('data:image/svg+xml') || imageUrl.includes('<svg') || imageUrl.endsWith('.svg')) {
-        return 'svg'
-    }
-
-    // Handle HTTP URLs by fetching and checking content-type
-    if (imageUrl.startsWith('http')) {
-        try {
-            const response = await fetch(imageUrl)
-            const contentType = response.headers.get('content-type')
-
-            if (contentType) {
-                if (contentType.includes('image/jpeg')) {
-                    return 'jpeg'
-                } else if (contentType.includes('image/png')) {
-                    return 'png'
-                } else if (contentType.includes('image/svg+xml')) {
-                    return 'svg'
-                }
-            }
-        } catch (error) {
-            console.error('Error fetching image type:', error)
-        }
-    }
-
-    return 'unknown'
 }
